Started PPU.
Cycle changed from u8 to u16 do  accommodate OAM DMA. that cycle for up
to 513.

Started an ASM module to run opcodes more directly.

+++ src/cpu/cycle.rs
@@ -3,15 +3,15 @@
 // For the actual cycle state see:
 // - http://www.visual6502.org/wiki/index.php?title=6502_Timing_States
 // - http://www.visual6502.org/wiki/index.php?title=6502_State_Machine
-pub const T1: u8 = 1;
-pub const T2: u8 = 2;
-pub const T3: u8 = 3;
-pub const T4: u8 = 4;
-pub const T5: u8 = 5;
-pub const T6: u8 = 6;
-pub const T7: u8 = 7;
-pub const T8: u8 = 8;
+pub const T1: u16 = 1;
+pub const T2: u16 = 2;
+pub const T3: u16 = 3;
+pub const T4: u16 = 4;
+pub const T5: u16 = 5;
+pub const T6: u16 = 6;
+pub const T7: u16 = 7;
+pub const T8: u16 = 8;
 
-pub const FIRST: u8 = T2;
-pub const LAST: u8 = T1;
-pub const NEX_TO_LAST: u8 = LAST - 1;
+pub const FIRST: u16 = T2;
+pub const LAST: u16 = T1;
+pub const NEX_TO_LAST: u16 = LAST - 1;
